Rework PageFinder restrictions
The idea behind this rework is that PageFinder should reject a page_id
if and only if it does not resolve to a sub-path of content_path.  This
changes the previous behavior in the following ways:

* non-ASCII characters are now allowed (e.g. "résumé" => "pages/résumé")

* page_ids with leading dots are now explicitly allowed (e.g.
  ".gitconfig" => "pages/.gitconfig") [NOTE: this could already be
  hacked into working with e.g. "_/../.file" => "pages/.file" or
  "subdir/.file" => "pages/subdir/.file"]

* page_ids which resolve to outside the content_path immediately error,
  instead of coercing to a sub-path (e.g. "../secret" => error, instead
  of "../secret" => "pages/secret")

* the trailing slash on content_path is now optional (previously,
  forgetting the trailing slash resulted in a confusing 404)

This rework also removes (obsoletes?) a subtle bug in the page_id
character whitelist wherein "+-=" was interpreted as ("+".."=") instead
of ["+", "-", "="].  Ironically, this bug is what allowed nested paths
in page_ids to work at all.