Utilize Rails view lookup paths to find templates
In the root_route constraint, make use of the Rails view context lookup
in order to determine if a given template exists or not. This leaves us
with not having to deal with manually looking files up in the view
structure since Rails is good at this already.

Doing this allows us to resolve 241 and 247